Replacement of itoa with SNPRINTF didn't work correctly with float arguments
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@910 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/UI/GuiHandler.cpp
@@ -898,7 +898,7 @@ void CGuiHandler::DrawFront(int button,float maxSize,float sizeDiv)
 
 	if(sizeDiv!=0){
 		char c[40];
-		SNPRINTF(c, 40, "%d",pos1.distance2D(pos2)/sizeDiv );
+		SNPRINTF(c, 40, "%d", (int)(pos1.distance2D(pos2)/sizeDiv) );
 		mouse->cursorTextRight=c;
 	}
 	glEnable(GL_FOG);
